Syslog
We're using a syslog and want to display the date, time, and process ID that's inside the square brackets. Read each line of the syslog and pass the contents to the show_time_of_pid function. Fill in the gaps to extract the date, time, and process id from the passed line, and return this format: Jul 6 14:01:23 pid:29440.